--- src/app/auth/sign-in/profile/_shared/services/type.ts
@@ -1,5 +1,0 @@
-export interface PatchMemberRequest {
-  nickname: string;
-  profileImg?: File;
-  introduction?: string;
-}
